Adding state trie logic
Ran an experiment which proved that newly created accounts do not get added to the state trie. Only accounts which have valid transactions against them are added to the state trie. The logic behind this is that creating a new account is free and transacting between accounts costs gas. This clever logic negates an attack vector.